Add mongo connection measurement tool

+++ Models/MongoSizeRecord.cs
@@ -5,8 +5,8 @@
         public string PodName { get; set; }
         public string Database { get; set; }
         public string Collection { get; set; }
-        public int DocumentSize { get; set; }
-        public int AverageDocumentSize { get; set; }
-        public int CollectionSize { get; set; }
+        public float DocumentSize { get; set; }
+        public float AverageDocumentSize { get; set; }
+        public float CollectionSize { get; set; }
     }
 }
